+++ Source/Scene/BingMapsImageryProvider.js
@@ -415,6 +415,20 @@ define([
             get : function() {
                 return this._credit;
             }
+        },
+
+        /**
+         * Gets a value indicating whether or not the images provided by this imagery provider
+         * include an alpha channel.  If this property is false, an alpha channel, if present, will
+         * be ignored.  If this property is true, any images without an alpha channel will be treated
+         * as if their alpha is 1.0 everywhere.  Setting this property to false reduces memory usage
+         * and texture upload time.
+         * @type {Boolean}
+         */
+        hasAlphaChannel : {
+            get : function() {
+                return false;
+            }
         }
     });